requirements.txt
pydantic
colorlog
pydantic_core
websockets
uvicorn
fastapi
ruamel.yaml
requests
aiosqlite
pillow
xpinyin
pytubefix
Beautifulsoup4
aiofiles
apscheduler
jmcomic
img2pdf
playwright
bilibili-api-python
numpy
pydub
websockets
openai
aiohttp
flask
flask-cors
cryptography
httpx<0.28.0
redis
jieba
scikit-learn
psutil
pyzipper
pycairo
watchdog